Rename license.md to LICENSE.md
The MIT license requires that the copyright notice in the license file be
distributed by works containing the package.

Normally, this type of thing is easily handled by distributing the
`vendor/` directory managed by `go mod vendor`.  However, `go mod vendor`

However, `go mod vendor` only looks for files with a static list of
filename prefixes; notably it looks for "LICENSE" but not "license", so
the `license.md` file wouldn't be included.

Have it be included by renaming it to `LICENSE.md`.

https://github.com/golang/go/blob/go1.17.3/src/cmd/go/internal/modcmd/vendor.go#L309-L337